+++ core/common/src/main/java/org/openengsb/core/common/taskbox/TaskboxService.java
@@ -14,11 +14,15 @@
  * limitations under the License.
  */
 
-package org.openengsb.core.taskbox;
+package org.openengsb.core.common.taskbox;
+
+import org.openengsb.core.common.Event;
+import org.openengsb.core.common.workflow.WorkflowException;
+import org.openengsb.core.common.taskbox.model.Task;
 
 /**
  * The Taskbox is a service which can be used when human interaction is required, e.g. by help desk applications. This
- * core part core is responsible for storing tasks, throwing events and starting workflows. Therefore it provides
+ * core part is responsible for storing tasks, throwing events and starting workflows. Therefore it provides
  * methods which can be called by workflows e.g. assigning a task to different user-roles (such as case worker or
  * developer) or setting a task status. Another job is to choose the right wicket panel from the UI project to display
  * the right information in a certain situation.
@@ -29,20 +33,33 @@ package org.openengsb.core.taskbox;
 public interface TaskboxService {
     /**
      * Gets the message set by a workflow
+     * Used for testing purposes up to now
      * 
      * @throws TaskboxException when the message is not set
      */
     String getWorkflowMessage() throws TaskboxException;
 
     /**
-     * Starts a test workflow
+     * Starts a test workflow 
+     * workflowName - the name of the workflow to be started
+     * taskVariableName - the name of the variable containing the taskObject in the workflow 
+     * task - the taskobject to be reasoned about
      * 
      * @throws TaskboxException when the test workflow could not be started
      */
-    void startWorkflow() throws TaskboxException;
+
+    void startWorkflow(String workflowName, String taskVariableName, Task task) throws TaskboxException;
 
     /**
      * Used by a workflow to set a message
+     * Used for testing purposes up to now
      */
     void setWorkflowMessage(String message);
+
+    /**
+     * Redirect events to workflowService
+     * 
+     * @throws WorkflowException
+     */
+    void processEvent(Event event) throws WorkflowException;
 }